+++ src/main/java/com/blue/view/MemberController.java
@@ -28,8 +28,10 @@ import org.springframework.web.bind.support.SessionStatus;
 import org.springframework.web.multipart.MultipartFile;
 import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
+import com.blue.dto.AlarmVO;
 import com.blue.dto.MemberVO;
 import com.blue.dto.PostVO;
+import com.blue.service.AlarmService;
 import com.blue.service.EmailService;
 import com.blue.service.MemberService;
 import com.blue.service.PostService;
@@ -46,6 +48,8 @@ public class MemberController {
 	private PostService postService;
 	@Autowired
 	private EmailService emailService;
+	@Autowired
+	private AlarmService alarmService;
 
 	// 회원가입 화면 표시
 	@RequestMapping("/join_view")
@@ -142,6 +146,59 @@ public class MemberController {
 		memberService.insertMember(vo);
 
 		return "login";
+	}
+	
+	@GetMapping(value="/editProfile")
+	public String editProfile(HttpSession session, Model model) {
+		
+		if(session.getAttribute("loginUser") == null) {
+			//System.out.println("세션값 없음");
+			model.addAttribute("message", "로그인을 해주세요");
+			return "login";
+		} else {
+			
+		List<PostVO> hottestFeed = postService.getHottestFeed();
+		String session_Id = ((MemberVO) session.getAttribute("loginUser")).getMember_Id();
+
+		// 알람 리스트를 담는 부분
+    	List<AlarmVO> alarmList = alarmService.getAllAlarm(session_Id);
+    	
+    	int alarmListSize = alarmList.size();
+    	
+    	// 알람의 종류를 파악하는 부분
+    	for(int j=0; j<alarmList.size(); j++) {
+    		int kind = alarmList.get(j).getKind();
+    		if(kind == 1) {
+    			alarmList.get(j).setMessage(alarmList.get(j).getFrom_Mem() + "님께서 회원님을 팔로우 <br>하였습니다.");
+    		}else if(kind == 2) {
+    			alarmList.get(j).setMessage(alarmList.get(j).getFrom_Mem() + "님께서 회원님의 게시글에 <br>좋아요를 눌렀습니다.");
+    		}else if(kind == 3) {
+    			alarmList.get(j).setMessage(alarmList.get(j).getFrom_Mem() + "님께서 회원님의 게시글에 <br>댓글을 달았습니다.");
+    		}else if(kind == 4) {
+    			alarmList.get(j).setMessage(alarmList.get(j).getFrom_Mem() + "님께서 회원님의 댓글에 <br>좋아요를 눌렀습니다.");
+    		}else if(kind == 5) {
+    			alarmList.get(j).setMessage("회원님께서 문의하신 질문에 <br>답글이 달렸습니다.");
+    		}
+    	}
+    	
+    	MemberVO loginUser = (MemberVO) session.getAttribute("loginUser");
+    	
+    	String email = loginUser.getMember_Email();
+		int atIndex = email.indexOf("@");
+		String email_Id = email.substring(0, atIndex);
+		String email_add = email.substring(atIndex + 1);
+		
+		
+		model.addAttribute("loginUser", loginUser);
+		model.addAttribute("member_Email", email_Id);
+		model.addAttribute("email_add", email_add);
+    	model.addAttribute("alarmList", alarmList);
+		model.addAttribute("alarmListSize", alarmListSize);
+		model.addAttribute("hottestFeed", hottestFeed);
+		
+		return "edit_profile";
+		
+		}
 	}
 
 	// 회원정보 수정
